image_classification/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
PyYAML>=6.0
